Add framerail hook to consolidate request context
